Document changes to the hashing scheme based on Solar Designer's comments.
The main change is from :

  bcrypt -> YHSM -> PBKDF2

to

  bcrypt -> PBKDF2 -> YHSM -> PBKDF2(iter=1)

Also, the credentials are not stored in the userdb database accessible to
the frontends but instead in a private "credential store" only available
to the authentication backends.

The frontends keep the bcrypt salt away from the authentication backends,
to further isolate which function (server or servers) can do what. This
change limits an intruder on the backends from offline attacking the bcrypt
hashes seen on the backends, since the attacker does not know the salt used
to compute them.